Applied XCode 8 recommended settings + fixed all new warnings
With XCode 8 we got many more warnings which this patch fixes, along with a bit clean up. Significant changes were made for performSelector: calls, which often gave warnings about unknown selectors and leaks (which is normal with ARC, but needed a workaround).

+++ backend/wbpublic/objimpl/db.query/db_query_Resultset.cpp
@@ -407,7 +407,7 @@ public:
   {
     if (column_by_name.find(column) != column_by_name.end())
     {
-      return grt::DoubleRef(recordset->getDouble((long)column_by_name[column]));
+      return grt::DoubleRef(recordset->getDouble((uint32_t)column_by_name[column]));
     }
     throw std::invalid_argument(base::strfmt("invalid column %s for resultset", column.c_str()).c_str());
     return grt::DoubleRef(0.0);
